Added Basic data
Added a couple of JPA objects: Author and Book
Started as basic POJO, then added JPA annotations for the Entity, and
the ID.
Added configuration to application.properties to enable the H2 console
as don't have an interface yet.

Add the Data files

Add the Data files described in last commit.

Added Basic data

Added a couple of JPA objects: Author and Book
Started as basic POJO, then added JPA annotations for the Entity, and
the ID.
Added configuration to application.properties to enable the H2 console
as don't have an interface yet.